Reestruturação das animações
Foi uma bagunça recriar essa classe de animaçoes. Colocar os tipos de animação e as instâncias de fato em listas e chamá-las na hora de rodar gerou uma grande confusão. Porém, está funcionando e eu quero registrar essa versão num commit pq achei mto interessante chegar nela e o resultado.

Agora está muito mais fácil rodar uma animação na tela (chama Animation.start passando o tipo e as variáveis).

Agora vou criar um Enum para os tipos de animações, para ficar mais organizado.

+++ src/utilities/TimeCounter.java
@@ -28,7 +28,7 @@ public class TimeCounter
 	
 	public double getCounter() { return counter ;}
 	public double getDuration() {return duration ;}	
-	public void setDuration(int duration) { this.duration = duration ;}
+	public void setDuration(double duration) { this.duration = duration ;}
 
 	public void start() { initialTime = System.nanoTime() * Math.pow(10, -9) ; active = true ;}
 	public void stop() { active = false ;}
